# src/routes/gatewayRoutes.ts
import { Router } from "express";
import { authenticateUser } from "@middlewares/authMiddleware";
import { UserType } from "@models/UserType";
import {
  createGateway,
  deleteGateway,
  getGatewaysByNetwork,
  getGatewayByMac,
  updateGateway
} from "@controllers/gatewayController";
import { GatewayFromJSON } from "@dto/Gateway";

const router = Router({ mergeParams: true });

// Get all gateways (Any authenticated user)
router.get("", authenticateUser(), async (req, res, next) => {
  try {
    res.status(200).json(await getGatewaysByNetwork(req.params.networkCode));
  } catch (error) {
    next(error);
  }
});

// Create a new gateway (Admin & Operator)
router.post("", authenticateUser([UserType.Admin, UserType.Operator]), async (req, res, next) => {
  try {
    await createGateway(req.params.networkCode, GatewayFromJSON(req.body));
    res.status(201).send();
  } catch (error) {
    next(error);
  }
});

// Get a specific gateway (Any authenticated user)
router.get("/:gatewayMac", authenticateUser(), async (req, res, next) => {
  try {
    res.status(200).json(await getGatewayByMac(req.params.gatewayMac, req.params.networkCode));
  } catch (error) {
    next(error);
  }
});

// Update a gateway (Admin & Operator)
router.patch("/:gatewayMac", authenticateUser([UserType.Admin, UserType.Operator]), async (req, res, next) => {
  try {
    await updateGateway(req.params.gatewayMac, req.body, req.params.networkCode);
    res.status(204).send();
  } catch (error) {
    next(error);
  }
});

// Delete a gateway (Admin & Operator)
router.delete("/:gatewayMac", authenticateUser([UserType.Admin, UserType.Operator]), async (req, res, next) => {
  try {
    await deleteGateway(req.params.gatewayMac, req.params.networkCode);
    res.status(204).send();
  } catch (error) {
    next(error);
  }
});

export default router;